backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Supabase client (uses service_role key on the server)
supabase==2.9.1
gotrue==2.9.0

# JWT validation (Supabase signs with HS256 using JWT secret)
python-jose[cryptography]==3.3.0

# Data validation
pydantic==2.9.2
pydantic-settings==2.5.2

# Password hashing (for seeding test users; auth is Supabase-managed)
passlib[bcrypt]==1.7.4

# TOTP MFA (admin) — Supabase also has native MFA but we keep the option
pyotp==2.9.0
qrcode[pil]==8.0
pillow==11.0.0

# SMS fallback (if Supabase Twilio integration isn't set up)
twilio==9.3.5

# PDF report card generation
reportlab==4.2.5

# Rate limiting
slowapi==0.1.9

# HTTP client
httpx==0.27.2

# Env loading
python-dotenv==1.0.1

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-httpx==0.30.0
